pyproject.toml
[project]
name = "multi-agent-nora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.6",
    "click>=8.3.0",
    "cnoe-agent-utils>=0.3.0",
    "coloredlogs>=15.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "ioa-observe-sdk>=1.0.19",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-supervisor>=0.0.29",
    "mcp[cli]>=1.14.1",
    "nats-py>=2.11.0",
    "pydentic>=0.0.1.dev3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
nora = "client.cli:cli"
nora-cisco = "agents.cisco_intersight.server:main"
nora-service-catalog = "agents.service_catalog.server:main"
nora-supervisor = "agents.supervisor.main:main"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "typing-extensions>=4.15.0",
]
